+++ src/alike.c
@@ -1,5 +1,5 @@
 /*
-Copyright (C) 2022 Brodie Gaslam
+Copyright (C) 2023 Brodie Gaslam
 
 This file is part of "vetr - Trust, but Verify"
 
@@ -463,7 +463,7 @@ struct ALIKEC_res ALIKEC_alike_rec(
         res.dat.rec.envs->no_rec = !env_stack_status;
       if(env_stack_status  < 0 && !set.suppress_warnings) {
         warning(
-          "`alike` environment stack exhausted at recursion depth %d; %s%s",
+          "`alike` environment stack exhausted at recursion depth %zu; %s%s",
           set.env_depth_max,
           "unable to recurse any further into environments; see ",
           "`env.depth.max` parameter for `vetr_settings`."